fix: put hooks and helpers under the same namespace
